Create iFrame scraper
This tool scrapes iframe from a webpage. Add it as a bookmark to your Bookmark bar then open the webpage you want to scrape. After that just click on this tool and it will scrape the iframe from that webpage. 

This tool helps find the video URL from a webpage.